requirements.txt
streamlit>=1.32
pandas>=2.2
python-dateutil>=2.8
plotly>=5.18
